backend/pyproject.toml
[project]
name = "leadlab-backend"
version = "0.1.0"
description = "LeadLab Starter Backend (FastAPI + SQLAlchemy)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "psycopg2-binary>=2.9",
  "python-dotenv>=1.0",
  "typer>=0.12"
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "httpx>=0.27", "ruff>=0.6"]

[project.scripts]
leadlab = "app.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
